+++ seedcase_sprout/core/sprout_checks/check_resource_path_string.py
